Abstract pre-task list routing to util function
As this routing has become quite complex, we abstract it to its own util for readability.
