+++ server/utils/applications/utils.test.ts
@@ -1,16 +1,24 @@
 import { isAfter } from 'date-fns'
+import { DeepMocked, createMock } from '@golevelup/ts-jest'
+import type { Request, Response } from 'express'
 import { applicationFactory, submittedApplicationFactory, timelineEventsFactory } from '../../testutils/factories'
 import {
   eligibilityQuestionIsAnswered,
   getApplicationTimelineEvents,
   getSideNavLinksForDocument,
   getSideNavLinksForApplication,
+  showMissingRequiredTasksOrTaskList,
 } from './utils'
+import { fetchErrorsAndUserInput } from '../validation'
 import { DateFormats } from '../dateUtils'
 import { getSections } from '../checkYourAnswersUtils'
 import config from '../../config'
+import paths from '../../paths/apply'
+import { TaskListService } from '../../services'
 
+jest.mock('../../services/taskListService')
 jest.mock('../checkYourAnswersUtils')
+jest.mock('../../utils/validation')
 
 const mockSections = [
   {
@@ -281,6 +289,176 @@ describe('utils', () => {
         { text: 'Task 1', href: '#task-1' },
         { text: 'Task 2', href: '#task-2' },
       ])
+    })
+  })
+
+  describe('showMissingRequiredTasksOrTaskList', () => {
+    const request: DeepMocked<Request> = createMock<Request>({ user: { token: 'SOME_TOKEN' } })
+    const response: DeepMocked<Response> = createMock<Response>({})
+
+    describe('when "Confirm eligibility" task is NOT complete', () => {
+      it('renders "Confirm eligibility" page from the "Before you start" section', async () => {
+        const application = applicationFactory.build({ data: {} })
+
+        const actual = showMissingRequiredTasksOrTaskList(request, response, application)
+
+        expect(actual).toEqual(
+          response.redirect(
+            paths.applications.pages.show({
+              id: application.id,
+              task: 'confirm-eligibility',
+              page: 'confirm-eligibility',
+            }),
+          ),
+        )
+      })
+    })
+
+    describe('when "Confirm eligibility" task is complete and the candidate is INELIGIBLE', () => {
+      it('renders the "ineligible" page', async () => {
+        const application = applicationFactory.build({
+          data: {
+            'confirm-eligibility': {
+              'confirm-eligibility': { isEligible: 'no' },
+            },
+          },
+        })
+
+        const actual = showMissingRequiredTasksOrTaskList(request, response, application)
+
+        expect(actual).toEqual(response.redirect(paths.applications.ineligible({ id: application.id })))
+      })
+    })
+
+    describe('when the person is confirmed ELIGIBLE but consent has been DENIED', () => {
+      it('redirects to the _consent refused_ page', async () => {
+        const application = applicationFactory.build({
+          data: {
+            'confirm-eligibility': {
+              'confirm-eligibility': { isEligible: 'yes' },
+            },
+            'confirm-consent': {
+              'confirm-consent': {
+                hasGivenConsent: 'no',
+                consentRefusalDetail: 'some reason',
+              },
+            },
+          },
+        })
+
+        const actual = showMissingRequiredTasksOrTaskList(request, response, application)
+
+        expect(actual).toEqual(response.redirect(paths.applications.consentRefused({ id: application.id })))
+      })
+    })
+
+    describe('when the person is confirmed ELIGIBLE but the consent task has not been completed', () => {
+      it('redirects to the _confirm consent_ page', async () => {
+        const application = applicationFactory.build({
+          data: {
+            'confirm-eligibility': {
+              'confirm-eligibility': { isEligible: 'yes' },
+            },
+          },
+        })
+
+        const actual = showMissingRequiredTasksOrTaskList(request, response, application)
+
+        expect(actual).toEqual(
+          response.redirect(
+            paths.applications.pages.show({
+              id: application.id,
+              task: 'confirm-consent',
+              page: 'confirm-consent',
+            }),
+          ),
+        )
+      })
+    })
+    describe('when the person is confirmed ELIGIBLE and consent is confirmed, but the HDC licence dates task has not been completed', () => {
+      it('redirects to the _HDC licence dates_ page', async () => {
+        const application = applicationFactory.build({
+          data: {
+            'confirm-eligibility': {
+              'confirm-eligibility': { isEligible: 'yes' },
+            },
+            'confirm-consent': {
+              'confirm-consent': {
+                hasGivenConsent: 'yes',
+                consentDate: '2022-02-22',
+                'consentDate-year': '2022',
+                'consentDate-month': '2',
+                'consentDate-day': '22',
+              },
+            },
+            'hdc-licence-dates': {
+              'hdc-licence-dates': {
+                hdcEligibilityDate: '2024-02-28',
+                'hdcEligibilityDate-year': '2024',
+                'hdcEligibilityDate-month': '2',
+                'hdcEligibilityDate-day': '28',
+                conditionalReleaseDate: '2024-02-22',
+                'conditionalReleaseDate-year': '2024',
+                'conditionalReleaseDate-month': '2',
+                'conditionalReleaseDate-day': '22',
+              },
+            },
+          },
+        })
+
+        const stubTaskList = jest.fn()
+        ;(TaskListService as jest.Mock).mockImplementation(() => {
+          return stubTaskList
+        })
+        ;(fetchErrorsAndUserInput as jest.Mock).mockImplementation(() => {
+          return { errors: {}, errorSummary: [], userInput: {} }
+        })
+
+        const actual = showMissingRequiredTasksOrTaskList(request, response, application)
+
+        expect(actual).toEqual(
+          response.render('applications/taskList', {
+            application,
+            taskList: stubTaskList,
+            errors: {},
+            errorSummary: [],
+            referrer: 'some-referrer/',
+          }),
+        )
+      })
+    })
+
+    describe('when eligibility and consent have been confirmed, and HDC dates have been entered', () => {
+      it('redirects to the _task list_ page', async () => {
+        const application = applicationFactory.build({
+          data: {
+            'confirm-eligibility': {
+              'confirm-eligibility': { isEligible: 'yes' },
+            },
+            'confirm-consent': {
+              'confirm-consent': {
+                hasGivenConsent: 'yes',
+                consentDate: '2022-02-22',
+                'consentDate-year': '2022',
+                'consentDate-month': '2',
+                'consentDate-day': '22',
+              },
+            },
+          },
+        })
+
+        const actual = showMissingRequiredTasksOrTaskList(request, response, application)
+
+        expect(actual).toEqual(
+          response.redirect(
+            paths.applications.pages.show({
+              id: application.id,
+              task: 'hdc-licence-dates',
+              page: 'hdc-licence-dates',
+            }),
+          ),
+        )
+      })
     })
   })
 })
